+++ jme3-terrain/src/main/java/com/jme3/terrain/heightmap/ParticleDepositionHeightMap.java
@@ -313,7 +313,7 @@ public class ParticleDepositionHeightMap extends AbstractHeightMap {
         //transfer the new terrain into the height map.
         for (int i = 0; i < size; i++) {
             for (int j = 0; j < size; j++) {
-                setHeightAtPoint((float) tempBuffer[i][j], j, i);
+                setHeightAtPoint(tempBuffer[i][j], j, i);
             }
         }
         erodeTerrain();
